Optionally cut the auth paths at a specific layer
This is the SPHINCS way to remove duplicates in the upper layers.
However, it looks like this will lead to worse results than keeping
the few nodes in the upper layers.
In most cases, the nodes in the upper layers are not part of the
combined auth path in the first place, since they can be reproduced
from lower layers.